Issue #23 - Quick refactor to allow import and reuse
This fixes the global access to args by added to PDFCreator class
for minimum_page_length.

NOTE PDFCreator() can be called externally but it expects an object
that looks like the argparse object returned for the global `parser`.
So either construct a custom object or call using code similar to:

    ...
    args = txt2pdf.parser.parse_args([source_file_name, '--encoding=' + source_encoding, '--output=' + pdf_file_name, '--media=letter', '--author=me', '--quiet'])
    PDFCreator(args, Margins(
       args.margin_right,
       args.margin_left,
       args.margin_top,
       args.margin_bottom)).generate()